Fixed routing for /Character/Dirichlet
I think that at some point someone moved the Dirichlet character routing
from website.py to DirichletCharacter.py. Except that they didn't remove
some routing from website.py, so the new routing wasn't used. Later on
someone came along and removed the extraneous routing in website.py, and
probably didn't check that the website worked after that, because the
routing in website.py was the only correct routing.

This is a small fix to make things work in DirichletCharacter.py.