Fixed Page loading problem #513
The real problem was that we were always looking up field labels which takes quite a while in general.
However, the existing ones were actually stored in the db and this is how it should be done.
We have to make sure that in the future this data gets updated but for now this fixes at least the loading issue.
Overall, this improves the performance of most modular forms pages a lot.