Check if cm_disc has been set to a valid value.
The method that sets the CM info currently fails sometimes because of
invalid data in the database, as it seems. I saw a few cases, e.g.
/ModularForm/GL2/Q/holomorphic/3/6/1/a/, where the information that this is
a CM form is corretly displayed but the CM field that is shown is
displayed as Q(\sqrt(0)). For now, I only check in the template that
the CM discriminant is a negative number.