call p2sage on root number, fix p2sage.
The plotting code was trying to multiply the values of the plot by the
root number, which was a string in some cases. Meanwhile, p2sage didn't
really do the right thing, because it was returning a list of strings
sometimes when the input was a list of strings. Additionally, it didn't
take into account that the input might expect some variables to be
defined.