Track number of coeffs of Maass form even when only one cusp is computed
I #2075, it was noted that some Maass forms are missing coefficients,
even though the associated L-function *does* have coefficients. This is
due to the fact that more data on some Maass forms is available than
some others. The number of coefficients is called when creating the
table displaying the coefficients, but this is only tracked correctly
when the database has lots of coefficients of the Maass form included.

This (small) change causes the number of coefficients to be tracked
always, allowing the table to be displayed.

As an example, this fixes the display of

    /ModularForm/GL2/Q/Maass/5411a477acf7560ce8a6f594

(though note that only 59 coefficients are currently available).

NOTE
----
  A long-term solution should include updating the database with further
  coefficients, including the coefficients at the different cusps.

RESOLVES
--------
  #2075