Introduce precision-cutoffs in Maass coefficients
In #2076, it was noted that imaginary parts with really small value,
like 10e-16, would look better if they were displayed as zero.

This commit introduces a precision (`prec`) argument in mwf_main which
makes this happen. Currently the precision is hard-coded to be 1e-9, as
suggested by David Farmer in #2076.

NOTE
----

A bit of refactoring is still necessary, and will be done before
submitting a pull request.