Add unittests for utils.py
This adds unittests for many of the utilities in utils.py. These are not
complete unittests, of course, but they are a start. This adds tests for
the behavior of the functions:

    an_list
    coeff_to_poly
    comma
    display_multiset
    list_to_latex_matrix
    pair2complex
    pol_to_html
    rgbtohex
    signtocolour
    splitcoeff
    to_dict
    truncatenumber
    web_latex_ideal_fact
    web_latex_split_on_pm
    web_latex_split_on_re
    web_latex_split_on
    web_latex

These are frontend utility functions. I have added documentation and
examples of these utilities within the utility file as well.

I should note that there are not
tests for the functions concerning ajax, SON, MongoDB pagination,
logging, plot encoding, or flask debugging. These are all end-to-end
utilities, and it may even be a good idea to separate these into another
utility file. Together, the untested utilities are:

    class LazyMongoDBPagination
    class LinkedList
    class AjaxPool
    class MongoDBPagination
    class LmfdbFormatter
    ajax_more
    ajax_result
    ajax_url
    attribute_value_counts
    cached
    debug
    encode_plot
    flash_error
    image_callback
    image_src
    len_val_fn
    make_logger
    orddict_to_strlist
    order_values
    random_object_from_collection
    random_value_from_collection