Correct double parentheses in bmf ideal display
We use web_latex_ideal_fact from utils.py for bmf pages. However, unlike
other instances, the bmf page wants to insert the latex'd version of the
factored ideal into a longer mathjax equation. So I added the ability to
not auto-enclose the output of web_latex and web_latex_ideal_fact from
utils.py in mathjax equation-parentheses.

This problem was noted in #2264