Create pages at /L/lhash/<Lhash>/
This is a progress update on having pages directly for each L-function.
At this point, there are some design choices that need to be made, which
is why this is a good place to stop and consider. I list the major ones
here.

With this, it is possible to go to, for example, the L-functions

    http://localhost:37777/L/lhash/1938322253992393114/
    http://localhost:37777/L/lhash/dirichlet_L_6253.458/

These pages are arrived at from the lhash and values stored in the
database. For initial comparisons, other places where this exact
L-function appear in the "Instances" box in the gaga box. [In the
future, we may want to have a cannonical page for each L-function, and
so this Instances box would need to be removed. I have indicated the
lines to remove in the code with `#TODO DLD` marks.] It is good to
compare the Lhash version against another instance, to see some
differences.

In the example above, there are a few obvious differences. Most of these
are because this "cannonical" Lhash page doesn't know what sort of
L-function it is, and so it omits things displayed using this
information.

Let's see some of these for the first example page above.

1. The "Related objects" box on the side doesn't appear.

  This is because the related objects (created in `lfunctions.main.py`)
  populated the topics on the gaga box differently based on the type of
  the L-function. But Lhash doesn't have a type.

2. The Euler Product is displayed incorrectly.

  The Euler product is essentially gathered in `lfuncEPtex` in
  `lfunctions.Lfunctionutilities.py`. This has different cases depending
  on the type of the L-function. So an elliptic curve Euler product is
  constructed differently than a Dirichlet L-function's Euler product,
  for instance.

  I note that the good and bad factors are stored in the database. So it is
  possible to rewrite the Euler product constructor to produce the
  correct output using only the contents of the database. There is a
  question of presentation, though. Do we want a Dirichlet L-function's
  Euler product to have the correct `chi` displayed, for instance?

3. The plot of the Z-function doesn't appear.

  The plot is presented by modifying the url of the L-function to arrive
  at the url for the plot. It is possible to still do this, and just
  handle urls for the plots under Lhash's, but as this is an initial set
  of changes I thought that handling the plot is less important.