Remove Lkey from lfunctions code
In #2298, we noticed that many (not all) of the Lfunction objects in the
code implement Lkey. But this is never used and no one seems to remember
what it was for. To prevent others in the future from losing time while
trying to figure out if these have a purpose, this commit removes all
instances of Lkey from the code.

Resolves #2298.