Add functional test for citations pages
The problem in #2332 went unnoticed because of a lack of testing. This
commit adds tests for the base citations page and the "How to Cite" the
lmfdb page. These tests really just check that these pages exist. It
also tests a specific reference in the citation list and tests for its
bib entry in the associated bibliography list. This would catch the
problem from #2332.

Note that I've also submitted a pull request to the LMFDB/citations
repository to prevent this issue from coming up. This is
LMFDB/citations#9