Remove Euler product from lhash page
The way we display the Euler product depends on data which is not stored
in the database, such as the `self.langlands` parameter or the type of
L-function.

This commit removes Euler products from lhash pages and instead refers
viewers to other instances of that L-function for more. It also defaults
`self.langlands` to False instead of True, but this only affects Euler
product display.