Allow filtering knowls by category and quality
This makes it possible to click on a category in /knowledge/, and then
choose some subset of qualities and filter them out. It does *not* allow
the other direction (unfortunately), and I struggled to find a
reasonable way to allow one to maintain filters when clicknig on another
category. It is possible if I were to do a larger rewrite, or if I were
to mix GET and POST requests beyond what I think is reasonable.