Fix test
The old test was for checking for another element in the Galois orbit,
but the Galois orbit computed wasn't ordered --- its order came from the
way in which it was computed on the fly. Now the Galois orbit presented
is ordered (from smallest Conrey index to highest), and so displaying
only the first several doesn't hit the same one.

Now I check for a different character in the Galois orbit.