Correct circular reference, show zeta value at 1/2
There is a circular reference preventing the sageLfunction from being
correctly created in the zeta function. This is caused from
general_webpage_info calling a sequence of functions that also called
the L-function constructor itself.

This causes the central value at 1/2 to not be computed, as noted in

On the other hand, we are moving away from using lcalc to compute any
values on the fly, so this is a temporary measure. But I chose to fix
the circular reference.