Make character notation in sidebar consistent
In #3207, Edgar pointed out that the way we notate Dirichlet and Artin
representations in the sidebar was inconsistent between beta and main.
This makes them consistent: on main, that section appears like

```
Representations:
  Dirichlet Characters
  Artin
```

and on beta, that section appears like

```
Representations:
  Characters:
    Dirichlet
    Hecke
  Galois:
    Artin
    mod l
```

Notes
-----

I might add that this is the first instance of something appearing on
main that does not appear on beta (I think) --- here, it's "Dirichlet
Characters" instead of "Dirichlet".

There is a small bit of if-elif-else logic added to the template. If
there comes a time when this is no longer needed, perhaps because the
"Character" and "Galois" headings are added permanently, then this can
be simplified to `if BETA or not pt2.status`, and anything regarding the
status "notonbeta" can be removed.