Update Dirichlet character search error message
As a result of previous Dirichlet character updates, the Dirichlet
character search-by-label-or-modulus search also allows searching by
Galois orbit label. This means that one could search for

- `13.2` (to get the character with Contry label 13.2),
- `13`   (to get the group of characters mod 13), or
- `13.f` (to get that particular Galois orbit).

I hadn't noticed this before. Prior to this commit, if you gave an
invalid orbit label, like 13.z, a flash message would instruct you to
replace z with a valid *integer* that's coprime to the modulus.

This commit updates the error message to include the possibility that
the user is searching by Galois orbit. It also states that one can
search by orbit label, and gives a knowl to the orbit label.

This commit resolves #3357.