Add sage code for dirchars without conrey package
This commit adds code for computing Dirichlet characters without using
the extra package conrey_dirichlet_characters from Bober's github. In
addition, this now also links to that package from the Dirichlet
character pages.

To see this code, go to a page such as

https://www.lmfdb.org/Character/Dirichlet/60/7

and click on the `Show commands for SageMath` option.

Right now, I present both the conrey-code and an independent sage code.
We could choose to only show one if desired.

There is a choice to be made for the code down the rest of the page.
Almost all of the code snippets work for both conrey_characters and for
sage_characters, except for the final couple (where we allow explicit
evaluation of Gauss sums, Jacobi sums, and Kloosterman sums for whatever
reason).