Add eta quotients for classical modular forms
This commit adds eta quotients to pages on classical modular newform
orbit pages, if they exist.

There are a few design decisions here, and I document my reasons for
them now. Firstly, this is a relatively rare property, and so it should
not be a column in the db associated to each modular form orbit --- it
would simply be "None" for most of them. Thus it should either be its
own table in the db, hardcoded in the file, or in a yaml. I have chosen
to put this in a yaml file, similar to how some parts of the LMFDB parse
a yaml file to generate their code segments.

This yaml file needs to be parsed sometime and somewhere. It could
either be parsed in web_newform.py or in main.py. I've chosen to parse
it as late as possible, exactly in the rendering function in main.py.
But this is more open to interpretation.

There remain a few things before fully live.
1. This makes the bottom_knowls for functions expressible as eta
   products obsolete. These should be removed so that the list of
   "interesting forms" isn't filled with them.
2.  This also reuses a knowl, mf.half_integral_weight.eta_quotients,
   which does not seem to have been revisited in a long time. It may
   need to be rewritten.

See also the discussion in #2783.