Allow multiplication and exponentiation in input boxes, 4268 and 4288
Modified search_parsing.py to allow inputs involving multiplication and exponentiation. This uses regex to verify the inputted text only uses [0-9], (),^,* symbols and then attempts a ZZ(gp()). If ZZ(gp()) succeeds, the search is run with that, otherwise an error is displayed.

This is a modification from the previous fix attempt which used sage_eval() instead of ZZ(gp()).

See issues: #4268 and #4288.