Warn about --restart
This is because the current state of pari threading makes running with
--restart segfaults immediately on clicking on `random elliptic curve over
Q` (throwing a long confusing stacktrace about cypari things). It
probably causes crashes on other pages that use pari to dynamically
compute data or plots.

This adds a small warning to the only place where users could learn
about the --restart option.